# docs/swagger.yaml
basePath: /api
definitions:
...
  common.ErrorDetailElement:
    properties:
      field:
        type: string
      message:
        type: string
      param:
        type: string
      struct_field:
        type: string
      tag:
        type: string
      value: {}
    type: object
  common.HttpError:
    properties:
      code:
        type: integer
      detail:
        items:
          $ref: '#/definitions/common.ErrorDetailElement'
        type: array
      message:
        type: string
    type: object
...
  models.Commodity:
    properties:
      category:
        type: string
      default_name:
        type: string
      id:
        type: integer
      produce_address:
        type: string
      produce_at:
        $ref: '#/definitions/models.MyTime'
    type: object
  models.CommodityItem:
    properties:
      commodity:
        $ref: '#/definitions/models.Commodity'
      commodity_id:
        type: integer
      id:
        type: integer
      item_name:
        type: string
      platform:
        $ref: '#/definitions/models.Platform'
      platform_id:
        type: integer
      price:
        type: number
      seller:
        $ref: '#/definitions/models.Seller'
      seller_id:
        type: integer
      update_at:
        $ref: '#/definitions/models.MyTime'
    type: object
...
  models.FavoriteStatisticsResponse:
    properties:
      commodityItem:
        $ref: '#/definitions/models.CommodityItem'
      count:
        type: integer
    type: object
...
  models.MyTime:
    properties:
      time.Time:
        type: string
    type: object
  models.Platform:
    properties:
      country:
        type: string
      id:
        type: integer
      name:
        type: string
      url:
        type: string
    type: object
  models.PriceChange:
    properties:
      commodity_item_id:
        type: integer
      commodityItem:
        $ref: '#/definitions/models.CommodityItem'
      id:
        type: integer
      new_price:
        type: number
      update_at:
        $ref: '#/definitions/models.MyTime'
    type: object
  models.PriceStatisticsResponse:
    properties:
      commodityItem:
        $ref: '#/definitions/models.CommodityItem'
      price_variance:
        type: number
    type: object
  models.Seller:
    properties:
      address:
        description: Age      int
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
  models.SummaryResponse:
    properties:
      category:
        type: string
      category_num:
        type: integer
      commodity:
        $ref: '#/definitions/models.Commodity'
      commodity_num:
        type: integer
      favorite_num:
        type: integer
      message_num:
        type: integer
      platform:
        $ref: '#/definitions/models.Platform'
      platform_num:
        type: integer
      price_change_num:
        type: integer
      seller:
        $ref: '#/definitions/models.Seller'
      seller_num:
        type: integer
    type: object
...
  priceChange.GetPriceChangeModel:
    properties:
      commodity_item_id:
        type: integer
      time_end:
        $ref: '#/definitions/models.MyTime'
      time_start:
        $ref: '#/definitions/models.MyTime'
    required:
    - commodity_item_id
    - time_end
    - time_start
    type: object
  priceChange.UpdatePriceChangeModel:
    properties:
      id:
        type: integer
      new_price:
        description: CommodityItemID int     `json:"commodity_item_id"`
        type: number
    type: object
  query.FavoriteStatisticsRequest:
    properties:
      age_end:
        type: integer
      age_start:
        type: integer
      all:
        type: boolean
      gender:
        type: boolean
    type: object
  query.PriceStatisticsRequest:
    properties:
      commodity_id:
        type: integer
      time_end:
        $ref: '#/definitions/models.MyTime'
      time_start:
        $ref: '#/definitions/models.MyTime'
    required:
    - commodity_id
    - time_end
    - time_start
    type: object
  seller.CreateSellerRequest:
    properties:
      address:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: polaris@fduhole.com
    name: Maintainer Polaris
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Database System Design Project
  version: 1.0.0
paths:
  /api/annual/summary:
    get:
      consumes:
      - application/json
      description: Get personal annual summary
      parameters:
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Get personal annual summary
      tags:
      - Query
  /api/annual/summary/all:
    get:
      consumes:
      - application/json
      description: Get all annual summary
      parameters:
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Get all annual summary
      tags:
      - Query
...
  /api/favorite/statistics:
    post:
      consumes:
      - application/json
      description: Get favorite statistics
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/query.FavoriteStatisticsRequest'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FavoriteStatisticsResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Get favorite statistics
      tags:
      - Query
...
  /api/item/history/{id}:
    delete:
      consumes:
      - application/json
      description: Delete batch priceChange by commodityItemID
      parameters:
      - description: commodityItem id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Delete batch priceChange by commodityItemID
      tags:
      - PriceChange
...
  /api/price/history:
    post:
      consumes:
      - application/json
      description: Get priceChange by ID
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/priceChange.GetPriceChangeModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PriceChange'
            type: array
      summary: Get priceChange by ID
      tags:
      - PriceChange
    put:
      consumes:
      - application/json
      description: Update priceChange
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/priceChange.UpdatePriceChangeModel'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Update priceChange
      tags:
      - PriceChange
  /api/price/history/{id}:
    delete:
      consumes:
      - application/json
      description: Delete priceChange by ID
      parameters:
      - description: priceChange id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Delete priceChange by ID
      tags:
      - PriceChange
  /api/price/history/batch:
    post:
      consumes:
      - application/json
      description: Add batch priceChange
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          items:
            $ref: '#/definitions/models.PriceChange'
          type: array
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Add batch priceChange
      tags:
      - PriceChange
...
  /api/price/statistics:
    post:
      consumes:
      - application/json
      description: Get price statistics
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/query.PriceStatisticsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PriceStatisticsResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Get price statistics
      tags:
      - Query
...
  /api/sellers:
    delete:
      consumes:
      - application/json
      description: Delete seller
      parameters:
      - description: seller id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Delete seller
      tags:
      - Seller
    post:
      consumes:
      - application/json
      description: Add seller
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/seller.CreateSellerRequest'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Add seller
      tags:
      - Seller
    put:
      consumes:
      - application/json
      description: Update seller
      parameters:
      - description: json
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.Seller'
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Update seller
      tags:
      - Seller
  /api/sellers/all:
    get:
      consumes:
      - application/json
      description: Get all sellers
      parameters:
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Seller'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Get all sellers
      tags:
      - Seller
  /api/sellers/data:
    get:
      consumes:
      - application/json
      description: Get seller info
      parameters:
      - description: Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.HttpError'
      summary: Get seller info
      tags:
      - Seller
...
swagger: "2.0"